+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/stat/config/PolarisStatProperties.java
@@ -47,7 +47,6 @@ public class PolarisStatProperties {
 	 */
 	private String path = "/metrics";
 
-
 	public boolean isEnabled() {
 		return enabled;
 	}