+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [Bugfix: update byte-buddy scope test to compile](https://github.com/Tencent/spring-cloud-tencent/pull/495)
 - [feature:add @ConditionalOnConfigReflectEnabled annotation](https://github.com/Tencent/spring-cloud-tencent/pull/496)
 - [Feature: zuul supports polaris router](https://github.com/Tencent/spring-cloud-tencent/pull/502)
+- [Feature：support pushGateway push metrics](https://github.com/Tencent/spring-cloud-tencent/pull/504)
 - [Refactor : optimize project and code](https://github.com/Tencent/spring-cloud-tencent/pull/506)
 - [Fix typo & Code optimization](https://github.com/Tencent/spring-cloud-tencent/pull/507)
 - [Code optimization for rpc-enhancement module](https://github.com/Tencent/spring-cloud-tencent/pull/510)